refactor: move webhook implementation from src/ to alc-core
Move ReqwestWebhookClient, PgWebhookService, fire_event_impl,
deliver_webhook, check_overdue_schedules and all 14 unit tests
into crates/alc-core/src/webhook.rs. src/webhook.rs is now a
thin re-export facade (4 lines).

src/ reduced from ~1,248 to ~637 lines total.
