+++ crates/alc-core/Cargo.toml
@@ -12,6 +12,7 @@ async-trait = "0.1"
 axum = "0.8"
 base64 = "0.22"
 chrono = { version = "0.4", features = ["serde"] }
+hex = "0.4"
 hmac = "0.12"
 jsonwebtoken = "9"
 reqwest = { version = "0.12", features = ["json"] }
@@ -21,7 +22,7 @@ serde_json = "1"
 sha2 = "0.10"
 sqlx = { version = "0.8", features = ["runtime-tokio", "tls-rustls", "postgres", "chrono", "uuid"] }
 thiserror = "2"
-tokio = { version = "1", features = ["sync"] }
+tokio = { version = "1", features = ["sync", "time"] }
 tracing = "0.1"
 uuid = { version = "1", features = ["v4", "serde"] }
 
@@ -29,6 +30,7 @@ uuid = { version = "1", features = ["v4", "serde"] }
 tokio = { version = "1", features = ["full", "test-util"] }
 wiremock = "0.6"
 rsa = "0.9"
+serde_json = "1"
 
 [features]
 test-helpers = []
